Conflicts:
	wa-apps/contacts/css/contacts.css
	wa-apps/contacts/js/compiled/contacts-jquery.js
	wa-apps/contacts/js/compiled/contacts.js
	wa-apps/contacts/js/fieldTypes.js
	wa-apps/contacts/js/wa.controller.js
	wa-apps/contacts/lib/actions/contacts/contactsContactsInfo.action.php
	wa-apps/contacts/lib/actions/contacts/contactsContactsMerge.controller.php
	wa-apps/contacts/lib/actions/contacts/contactsContactsMergeSelectMaster.action.php
	wa-apps/contacts/lib/config/app.php
	wa-apps/contacts/locale/ru_RU/LC_MESSAGES/contacts.mo
	wa-apps/contacts/locale/ru_RU/LC_MESSAGES/contacts.po
	wa-apps/contacts/templates/actions/contacts/ContactsInfo.html
	wa-apps/contacts/templates/actions/contacts/ContactsInfoTabs.html
	wa-apps/contacts/templates/actions/contacts/ContactsMergeSelectMaster.html
	wa-apps/site/locale/en_US/LC_MESSAGES/site.mo
	wa-system/contact/waContactPhoneField.class.php
	wa-system/webasyst/lib/cli/webasystCreateApp.cli.php
	wa-system/webasyst/lib/cli/webasystCreatePlugin.cli.php

+++ wa-system/contact/waContactFields.class.php
@@ -282,10 +282,10 @@ class waContactFields
 
         // Remove field from order lists first
         if (isset(self::$companyFields[$id])) {
-            self::disableField($id, 'company');
+            self::disableField($id, 'company', true);
         }
         if (isset(self::$personFields[$id])) {
-            self::disableField($id, 'person');
+            self::disableField($id, 'person', true);
         }
 
         $file = wa()->getConfig()->getConfigPath('custom_fields.php', true, 'contacts');
@@ -311,6 +311,14 @@ class waContactFields
             throw new waException('Unable to find field '.$id.' in '.$file);
         }
         unset($fields[$k]);
+        
+        $fields = array_values($fields);
+        foreach ($fields as $field) {
+            if ($field instanceof waContactField) {
+                $field->prepareVarExport();
+            }
+        }
+        
         waUtils::varExportToFile(array_values($fields), $file, true);
         unset(self::$fieldStatus[$id], self::$personDisabled[$id], self::$companyDisabled[$id]);
     }
@@ -351,6 +359,11 @@ class waContactFields
         }
         if (!$changed) {
             $fields[] = $field;
+        }
+        foreach ($fields as $field) {
+            if ($field instanceof waContactField) {
+                $field->prepareVarExport();
+            }
         }
         waUtils::varExportToFile($fields, $file, true);
 
@@ -457,9 +470,10 @@ class waContactFields
      * Remove given field from person or company order list.
      * @param $type string person|company
      * @param $id waContactField|int field ID or field instance.
+     * @param boolean $delete delete values from db or not
      * @throws waException
      */
-    public static function disableField($id, $type) {
+    public static function disableField($id, $type, $delete = false) {
         self::ensureStaticVars();
         if (is_object($id) && $id instanceof waContactField) {
             $id = $id->getId();
@@ -495,8 +509,10 @@ class waContactFields
          * @var waContactField $f
          */
 
-        // Remove data from DB
-        $f->getStorage()->deleteAll($id, $type);
+        if ($delete) {
+            // Remove data from DB
+            $f->getStorage()->deleteAll($id, $type);
+        }
 
         // Remove field from order file
         if (!is_readable($file)) {
